Serve node callbacks the storage-backed scope on every PHP version
FiberNodeScopeResolver's only remaining behavior - unwrap gatherers,
skip noop callbacks, hand rules the FiberScope - moves into the base
resolver, and the class goes together with the FnsrExtension autowiring
switch, the PHPSTAN_FNSR toggle, and the PHP 8.1 gate: nothing here
needs fibers anymore, so 7.4 and 8.0 get the same storage-backed rule
scope with walk-position answers as everything else. Test expectations
that keyed on the resolver split become unconditional.

(cherry picked from commit f5fede831af5dc8037bf93d034c0af29f2d37a3e)
