+++ src/Analyser/NodeScopeResolver.php
@@ -121,6 +121,8 @@ class NodeScopeResolver
 	/** @var array<string, true> filePath(string) => bool(true) */
 	private array $analysedFiles = [];
 
+	private ?ExpressionResultStorageStack $expressionResultStorageStack = null;
+
 	/**
 	 * @param ExtensionsCollection<FunctionParameterOutTypeExtension> $functionParameterOutTypeExtensions
 	 * @param ExtensionsCollection<MethodParameterOutTypeExtension> $methodParameterOutTypeExtensions
@@ -286,6 +288,12 @@ class NodeScopeResolver
 
 	public function storeExpressionResult(ExpressionResultStorage $storage, Expr $expr, ExpressionResult $expressionResult): void
 	{
+		// The storage only ever answers type questions from FiberScope, which
+		// resolves them from the before-scope. Storing just the before-scope
+		// keeps the storage from pinning throw points, impure points, scope
+		// callbacks and the after-scope of every expression until the end of
+		// the file.
+		$storage->storeBeforeScope($expr, $expressionResult->getBeforeScope());
 	}
 
 	/**
@@ -927,7 +935,37 @@ class NodeScopeResolver
 		ExpressionResultStorage $storage,
 	): void
 	{
-		$nodeCallback($node, $scope);
+		// Engine-feeding gatherers must observe the node at the emission
+		// position - their arrays are read as soon as the enclosing body walk
+		// returns. Gatherers are engine code and never ask about types -
+		// handing them the raw scope skips a FiberScope construction per
+		// emission; the scopes they capture (return statements, impure points)
+		// answer later asks through the storage hub like any MutatingScope.
+		while ($nodeCallback instanceof GatheringNodeCallback) {
+			($nodeCallback->getGatherer())($node, $scope);
+			$nodeCallback = $nodeCallback->getInner();
+		}
+
+		if ($nodeCallback instanceof NoopNodeCallback) {
+			return;
+		}
+
+		// post-order emission means the node's own result and every subnode
+		// result are already stored when the callback fires - FiberScope
+		// answers every ask synchronously from the storage; the emitting
+		// walk's storage is bound for the duration of the callback
+		$stack = $this->getExpressionResultStorageStack();
+		$stack->push($storage);
+		try {
+			$nodeCallback($node, $scope->toFiberScope());
+		} finally {
+			$stack->pop();
+		}
+	}
+
+	private function getExpressionResultStorageStack(): ExpressionResultStorageStack
+	{
+		return $this->expressionResultStorageStack ??= $this->container->getByType(ExpressionResultStorageStack::class);
 	}
 
 	/**
